Remove internal folder. Minor fixes
