+++ connectors/bigtable/flink-connector-gcp-bigtable/src/main/java/com/google/flink/connector/gcp/bigtable/utils/BigtableUtils.java
@@ -16,19 +16,21 @@
  * limitations under the License.
  */
 
-package com.google.flink.connector.gcp.bigtable.internal.serializers;
+package com.google.flink.connector.gcp.bigtable.utils;
 
 import org.apache.flink.api.connector.sink2.SinkWriter;
 
-import com.google.cloud.bigtable.data.v2.models.RowMutationEntry;
+import java.time.Instant;
 
-import java.io.Serializable;
+/** Collection of utilities for Bigtable Connector. */
+public class BigtableUtils {
+    public static final String DEFAULT_COLUMN_FAMILY = "flink";
 
-/**
- * Base serializer class.
- *
- * <p>This class implements the interface to serialize to {@link RowMutationEntry}.
- */
-public interface BaseRowMutationSerializer<T> extends Serializable {
-    RowMutationEntry serialize(T element, SinkWriter.Context context);
+    public static Long getTimestamp(SinkWriter.Context context) {
+        if ((context != null) && (context.timestamp() != null) && (context.timestamp() > 0)) {
+            return context.timestamp() * 1000;
+        } else {
+            return Instant.now().toEpochMilli() * 1000;
+        }
+    }
 }
